refactor(#3276): decompose compilePropertyAccess slice 3 — extract 4 more guard bands
Wave B (subtask of #3182), stacked on slice 2. Extract 4 more cohesive
early-return guard bands into property-access-dispatch.ts:
tryPrototypeMethodAndArityReads, tryLengthAndNameReads,
tryNamespaceConstantAndSymbolReads, tryStringLengthIteratorAndExternClassReads.

property-access.ts: 6473 -> 5573 LOC (-900; cumulative 7989 -> 5573 = -2416).
compilePropertyAccess itself ~3335 -> ~900 LOC (only preamble + 15 call blocks +
terminal struct-name block remain inline).
Byte-identity: IDENTICAL 39/39 (gc/standalone/wasi). tsc 0. Smoke 13/13.

✓
