+++ src/codegen/property-access.ts
@@ -199,10 +199,14 @@ import {
   tryDynamicReceiverRuntimeDispatchReads,
   tryGlobalThisAndProcessRead,
   tryIdentifierNamespaceAndStaticReceiverRead,
+  tryLengthAndNameReads,
+  tryNamespaceConstantAndSymbolReads,
   tryNativeErrorMemberRead,
   tryPinnedAndDeleteAwareDynamicGet,
   tryPrivateIdentifierRead,
+  tryPrototypeMethodAndArityReads,
   tryStandaloneBuiltinAndWasiMemberReads,
+  tryStringLengthIteratorAndExternClassReads,
   trySuperAndImportMetaRead,
 } from "./property-access-dispatch.js"; // (#3276) Wave B — extracted guard bands
 
@@ -341,7 +345,7 @@ export function resolveInheritedStaticProp(
  * optional parentheses around it). Used by NamedEvaluation to decide whether
  * a binding name is assigned to the function's .name. (#1049)
  */
-function isAnonymousFunctionDefinition(expr: ts.Expression): boolean {
+export function isAnonymousFunctionDefinition(expr: ts.Expression): boolean {
   while (ts.isParenthesizedExpression(expr)) expr = expr.expression;
   if (ts.isFunctionExpression(expr) && !expr.name) return true;
   if (ts.isArrowFunction(expr)) return true;
@@ -361,7 +365,7 @@ function isAnonymousFunctionDefinition(expr: ts.Expression): boolean {
  * `*-init-fn-name-class` whose `xCls2 = class { static name() {} }` asserts
  * `xCls2.name !== 'xCls2'`).
  */
-function classExpressionDefinesOwnName(expr: ts.Expression): boolean {
+export function classExpressionDefinesOwnName(expr: ts.Expression): boolean {
   let e: ts.Expression = expr;
   while (ts.isParenthesizedExpression(e)) e = e.expression;
   if (!ts.isClassExpression(e)) return false;
@@ -559,7 +563,7 @@ function isDirectBindCall(expr: ts.Expression): boolean {
  * direct check; this helper extends that to the deferred-storage form, which is
  * the bulk of the `built-ins/Function/prototype/bind/*` test262 cluster.
  */
-function isBindResultExpr(ctx: CodegenContext, expr: ts.Expression): boolean {
+export function isBindResultExpr(ctx: CodegenContext, expr: ts.Expression): boolean {
   if (isDirectBindCall(expr)) return true;
   if (ts.isIdentifier(expr)) {
     const sym = ctx.checker.getSymbolAtLocation(expr);
@@ -1946,7 +1950,7 @@ export function receiverMayBeNativeStringAtRuntime(ctx: CodegenContext, recv: ts
  * array-in-`any`, or `i32.const 0`). The builder receives the externref temp's
  * local index so it can re-push the original externref. Both arms produce i32.
  */
-function emitGuardedNativeStringLength(
+export function emitGuardedNativeStringLength(
   ctx: CodegenContext,
   fctx: FunctionContext,
   buildElseInstrs: (recvExternLocal: number) => Instr[],
@@ -2656,928 +2660,24 @@ export function compilePropertyAccess(
     if (__r !== PA_FALLTHROUGH) return __r;
   }
 
-  // (#1394) `ClassName.prototype.<method>` — emit a cached singleton
-  // closure-struct externref. The previous PR #294 emitted a fresh
-  // closure on every access, breaking the `c.m === C.prototype.m`
-  // identity assertion that 478 class/elements tests verify. The cache
-  // (one externref global per `${className}_${methodName}`, lazily
-  // initialised on first access) gives stable identity AND restores
-  // the +120 wins on instance-method-via-prototype yield-star
-  // extractions that PR #305's revert lost.
-  if (
-    ts.isPropertyAccessExpression(expr.expression) &&
-    ts.isIdentifier(expr.expression.expression) &&
-    expr.expression.name.text === "prototype"
-  ) {
-    const rawName = expr.expression.expression.text;
-    const className = ctx.classExprNameMap.get(rawName) ?? rawName;
-    if (ctx.classSet.has(className)) {
-      const fullName = `${className}_${propName}`;
-      // Only intercept actual instance methods. Skip static methods
-      // (they live on the constructor, not the prototype) and
-      // accessors (handled by the existing accessor path below).
-      if (ctx.classMethodSet.has(fullName) && !ctx.staticMethodSet.has(fullName)) {
-        const funcIdx = ctx.funcMap.get(classMemberFuncKey(ctx, fullName));
-        const structTypeIdx = ctx.structMap.get(className);
-        if (funcIdx !== undefined && structTypeIdx !== undefined) {
-          if (emitCachedMethodClosureAccess(ctx, fctx, fullName, funcIdx, structTypeIdx)) {
-            return { kind: "externref" };
-          }
-        }
-      }
-    }
+  {
+    const __r = tryPrototypeMethodAndArityReads(ctx, fctx, expr, propName, objType);
+    if (__r !== PA_FALLTHROUGH) return __r;
   }
 
-  // Handle Math.<method>.length — static function arity
-  if (
-    propName === "length" &&
-    ts.isPropertyAccessExpression(expr.expression) &&
-    ts.isIdentifier(expr.expression.expression) &&
-    expr.expression.expression.text === "Math"
-  ) {
-    const mathMethodArity: Record<string, number> = {
-      abs: 1,
-      ceil: 1,
-      floor: 1,
-      round: 1,
-      trunc: 1,
-      sign: 1,
-      sqrt: 1,
-      cbrt: 1,
-      clz32: 1,
-      fround: 1,
-      exp: 1,
-      expm1: 1,
-      log: 1,
-      log2: 1,
-      log10: 1,
-      log1p: 1,
-      sin: 1,
-      cos: 1,
-      tan: 1,
-      asin: 1,
-      acos: 1,
-      atan: 1,
-      sinh: 1,
-      cosh: 1,
-      tanh: 1,
-      asinh: 1,
-      acosh: 1,
-      atanh: 1,
-      min: 2,
-      max: 2,
-      pow: 2,
-      atan2: 2,
-      imul: 2,
-      hypot: 2,
-      random: 0,
-    };
-    const method = expr.expression.name.text;
-    if (method in mathMethodArity) {
-      fctx.body.push({ op: "f64.const", value: mathMethodArity[method]! });
-      return { kind: "f64" };
-    }
+  {
+    const __r = tryLengthAndNameReads(ctx, fctx, expr, propName, objType);
+    if (__r !== PA_FALLTHROUGH) return __r;
   }
 
-  // (#2187) `.length` on an `any`-typed identifier whose compiled local ValType
-  // is a native-string ref (e.g. a for-of var from a string-yielding generator).
-  // Must run BEFORE the Function/vec `.length` arms below: the static type is
-  // `any`, so those arms either miss or fall through to `__extern_length` (0
-  // standalone). At the VALUE level the receiver IS a string — read `len` (field
-  // 0 of `$AnyString`) natively. Tightly gated by `receiverIsNativeStringValType`
-  // (standalone + bare-identifier + any/unknown TS type + string-ref local).
-  if (propName === "length" && receiverIsNativeStringValType(ctx, fctx, expr.expression)) {
-    const recvType = compileExpression(ctx, fctx, expr.expression);
-    if (recvType && recvType.kind === "externref") {
-      coerceType(ctx, fctx, recvType, { kind: "ref_null", typeIdx: ctx.anyStrTypeIdx });
-    }
-    fctx.body.push({ op: "struct.get", typeIdx: ctx.anyStrTypeIdx, fieldIdx: 0 });
-    return { kind: "i32" };
+  {
+    const __r = tryNamespaceConstantAndSymbolReads(ctx, fctx, expr, propName, objType);
+    if (__r !== PA_FALLTHROUGH) return __r;
   }
 
-  // Handle Function.length — return the number of formal parameters
-  if (propName === "length") {
-    // (#1632a) `.length` on the result of `.bind(...)` must NOT be statically
-    // resolved to the target's param count — per spec it's
-    // `max(0, target.length - boundArgs.length)`. Fall through to the
-    // externref / __extern_get path so the host-bound function's actual
-    // `.length` is read.
-    const isBindResult = isBindResultExpr(ctx, expr.expression);
-    const callSigs = objType.getCallSignatures?.();
-    const constructSigs2 = objType.getConstructSignatures?.();
-    const lengthSigs =
-      callSigs && callSigs.length > 0 ? callSigs : constructSigs2 && constructSigs2.length > 0 ? constructSigs2 : null;
-    if (!isBindResult && lengthSigs && lengthSigs.length > 0) {
-      // For library/ambient functions, TS's param count can disagree with the
-      // runtime Function.length — the ES spec pins .length for methods like
-      // Array.prototype.toSorted to 1 even though the lib d.ts declares
-      // compareFn as optional ("?"). Defer to __extern_get("length") so the
-      // runtime value wins — but only when the root identifier of the chain
-      // is a known reachable global (BUILTIN_CTOR_NAMES / globalThis). Bare
-      // lib identifiers like `encodeURIComponent` or `DisposableStack` don't
-      // have a runtime externref binding, so __extern_get would throw.
-      const isLibrarySig = lengthSigs.some((s) => {
-        const decl = s.getDeclaration?.();
-        return decl?.getSourceFile().isDeclarationFile === true;
-      });
-      const BUILTIN_GLOBAL_ROOTS = new Set([
-        "Object",
-        "Array",
-        "Function",
-        "Symbol",
-        "Proxy",
-        "Reflect",
-        "Math",
-        "BigInt",
-        "JSON",
-        "Date",
-        "RegExp",
-        "ArrayBuffer",
-        "SharedArrayBuffer",
-        "DataView",
-        "Promise",
-        "WeakMap",
-        "WeakSet",
-        "WeakRef",
-        "FinalizationRegistry",
-        "Atomics",
-        "Iterator",
-        "Map",
-        "Set",
-        "Error",
-        "TypeError",
-        "RangeError",
-        "SyntaxError",
-        "URIError",
-        "EvalError",
-        "ReferenceError",
-        "String",
-        "Number",
-        "Boolean",
-        "Int8Array",
-        "Uint8Array",
-        "Uint8ClampedArray",
-        "Int16Array",
-        "Uint16Array",
-        "Int32Array",
-        "Uint32Array",
-        "Float32Array",
-        "Float64Array",
-        "BigInt64Array",
-        "BigUint64Array",
-        "globalThis",
-      ]);
-      let rootNode: ts.Expression = expr.expression;
-      while (ts.isPropertyAccessExpression(rootNode) || ts.isElementAccessExpression(rootNode)) {
-        rootNode = rootNode.expression;
-      }
-      const rootIsReachableBuiltin =
-        ts.isIdentifier(rootNode) &&
-        BUILTIN_GLOBAL_ROOTS.has(rootNode.text) &&
-        !fctx.localMap.has(rootNode.text) &&
-        !(fctx.boxedCaptures?.has(rootNode.text) ?? false);
-      if (!isLibrarySig || !rootIsReachableBuiltin) {
-        // ES spec: Function.length = number of required params before first
-        // optional/default/rest. TS forbids required-after-optional, so filtering
-        // out optional/default/rest is equivalent to iterating until the first one.
-        const sig = lengthSigs[0]!;
-        const paramCount = sig.parameters.filter((p: any) => {
-          const decl = p.valueDeclaration;
-          if (!decl || !ts.isParameter(decl)) return true;
-          return !decl.dotDotDotToken && !decl.questionToken && !decl.initializer;
-        }).length;
-        fctx.body.push({ op: "f64.const", value: paramCount });
-        return { kind: "f64" };
-      }
-      // Library signature rooted at a reachable builtin → fall through to
-      // externref / __extern_get path below.
-    }
-  }
-
-  // Handle Function.name — return the function name as a string
-  if (propName === "name") {
-    // (#1632a) `.name` on the result of `.bind(...)` must NOT be statically
-    // resolved to the target's symbol name — per spec it's `"bound " +
-    // target.name`. Fall through to the runtime __extern_get path so the
-    // host-bound function's actual `.name` property is read.
-    if (isBindResultExpr(ctx, expr.expression)) {
-      // Skip the static peephole entirely; fall through to the externref
-      // property-access path below. Covers both `fn.bind(...).name` and the
-      // deferred `const g = fn.bind(...); g.name` form (#1337).
-    } else {
-      const callSigs = objType.getCallSignatures?.();
-      const constructSigs = objType.getConstructSignatures?.();
-      const hasFuncSig = (callSigs && callSigs.length > 0) || (constructSigs && constructSigs.length > 0);
-      // (#1450) Even when the static type lacks call/construct signatures
-      // (catch parameter `any`, destructuring assignment target widened to
-      // contextual type, etc.), spec NamedEvaluation still applies if the
-      // identifier's binding declaration has an anonymous-fn / named-fn /
-      // class initializer. Pre-resolve here so destructuring patterns like
-      //   try {} catch ([fn = function(){}]) { fn.name }
-      // fold to the binding identifier text instead of the externref miss.
-      if (!hasFuncSig && ts.isIdentifier(expr.expression)) {
-        const sym = ctx.checker.getSymbolAtLocation(expr.expression);
-        const decl = sym?.valueDeclaration;
-        if (decl && (ts.isBindingElement(decl) || ts.isVariableDeclaration(decl))) {
-          let resolvedName: string | undefined;
-          if (decl.initializer) {
-            let initExpr: ts.Expression = decl.initializer;
-            while (ts.isParenthesizedExpression(initExpr)) initExpr = initExpr.expression;
-            if (isAnonymousFunctionDefinition(decl.initializer) && !classExpressionDefinesOwnName(decl.initializer)) {
-              // SingleNameBinding NamedEvaluation: anonymous fn/class inherits
-              // the binding identifier's text as its .name. (#2756) A class with
-              // its own `static name` member overrides the binding name, so skip
-              // synthesis there and fall through to the real property read.
-              resolvedName = expr.expression.text;
-            } else if (ts.isFunctionExpression(initExpr) && initExpr.name) {
-              // Named function expression keeps its own name (the binding
-              // identifier is ignored per spec).
-              resolvedName = initExpr.name.text;
-            } else if (ts.isClassExpression(initExpr) && initExpr.name) {
-              resolvedName = initExpr.name.text;
-            }
-          }
-          // (#2201) The binding initializer is not itself a function (e.g.
-          // `var value = 1` or no initializer), but a later logical-assignment
-          // may install an anonymous fn/class whose .name is NamedEvaluation'd
-          // to the LHS identifier.
-          if (resolvedName === undefined && sym) {
-            resolvedName = resolveLogicalAssignmentName(ctx, expr.expression, sym);
-          }
-          if (resolvedName !== undefined) {
-            addStringConstantGlobal(ctx, resolvedName);
-            return compileStringLiteral(ctx, fctx, resolvedName);
-          }
-        }
-      }
-      if (hasFuncSig) {
-        // Resolve the function name from the type symbol or the expression
-        let funcName = objType.getSymbol()?.name ?? "";
-        // __type, __function, __class, __object are anonymous type names from TS checker
-        if (funcName === "__type" || funcName === "__function" || funcName === "__class" || funcName === "__object")
-          funcName = "";
-        // Built-in globals declared as `declare var X: XConstructor` expose the
-        // interface name ("ArrayConstructor") as the type symbol, but the JS
-        // runtime `.name` is the declared identifier ("Array"). Strip the
-        // "Constructor" suffix when it matches the identifier text.
-        if (
-          funcName.endsWith("Constructor") &&
-          ts.isIdentifier(expr.expression) &&
-          expr.expression.text + "Constructor" === funcName
-        ) {
-          funcName = expr.expression.text;
-        }
-        // If the symbol name is empty (anonymous function), infer from context:
-        if (funcName === "") {
-          if (ts.isIdentifier(expr.expression)) {
-            // Direct variable access: f.name => infer "f"
-            // BUT: per ES spec (NamedEvaluation / IsAnonymousFunctionDefinition),
-            // if the binding initializer is a "covered" form like `(0, function(){})`
-            // (comma expression, call, etc.), the function's .name is NOT set to
-            // the binding name. Only direct FunctionExpression/ArrowFunction/
-            // ClassExpression (optionally parenthesized) qualifies. (#1049)
-            const sym = ctx.checker.getSymbolAtLocation(expr.expression);
-            const decl = sym?.valueDeclaration;
-            let initExpr: ts.Expression | undefined;
-            if (decl && (ts.isBindingElement(decl) || ts.isVariableDeclaration(decl)) && decl.initializer) {
-              initExpr = decl.initializer;
-            }
-            if (
-              initExpr !== undefined &&
-              (!isAnonymousFunctionDefinition(initExpr) || classExpressionDefinesOwnName(initExpr))
-            ) {
-              // Covered form — .name is "" (or whatever the inner fn already has).
-              // (#2756) A class with its own `static name` member overrides the
-              // NamedEvaluation binding name, so it is NOT the binding text either.
-              addStringConstantGlobal(ctx, "");
-              return compileStringLiteral(ctx, fctx, "");
-            }
-            funcName = expr.expression.text;
-          } else if (ts.isPropertyAccessExpression(expr.expression)) {
-            // Property access: obj.method.name => infer "method"
-            funcName = expr.expression.name.text;
-          } else if (
-            ts.isElementAccessExpression(expr.expression) &&
-            ts.isStringLiteral(expr.expression.argumentExpression)
-          ) {
-            // Element access: obj["method"].name => infer "method"
-            funcName = expr.expression.argumentExpression.text;
-          }
-        }
-        // Ensure the string constant is registered before compiling
-        addStringConstantGlobal(ctx, funcName);
-        return compileStringLiteral(ctx, fctx, funcName);
-      }
-    } // close `else` branch of the #1632a bind-result guard
-  }
-
-  // Handle array.length (vec struct: field 0 is the logical length)
-  if (propName === "length") {
-    // (#1742) `this.length` where `this` is the host-supplied `__current_this`
-    // externref but may carry a compiled vec at runtime (a closure body dispatched
-    // via `__call_fn_method_N`). The override `this` is typically `any` → externref,
-    // so the vec fast paths below never fire; without this guard the read falls
-    // through to `__extern_length`, which returns 0 for an externref-wrapped vec.
-    // Runtime `ref.test` against the registered vec types reads field 0 on a hit,
-    // `__extern_length` for a genuine host receiver. No-op otherwise.
-    {
-      // Only vec types are valid `.length` receivers (length at struct field 0);
-      // a non-vec static struct must NOT be read as a vec here.
-      const allTargets = thisReceiverGuardTargets(ctx, fctx, expr.expression, "element");
-      const targets = allTargets?.filter((idx) => {
-        const def = ctx.mod.types[idx];
-        return def?.kind === "struct" && def.fields[0]?.name === "length" && def.fields[1]?.name === "data";
-      });
-      if (targets !== undefined && targets.length > 0) {
-        const lenType: ValType = ctx.fast ? { kind: "i32" } : { kind: "f64" };
-        compileExpression(ctx, fctx, expr.expression); // → externref `this`
-        emitThisReceiverGuardConvert(
-          ctx,
-          fctx,
-          targets,
-          lenType,
-          (concreteType) => {
-            // [(ref $vec)] → length (vec struct field 0). Every registered vec
-            // type has `length` at field 0, so the matched concrete type works.
-            const vecIdx = (concreteType as { typeIdx: number }).typeIdx;
-            fctx.body.push({ op: "struct.get", typeIdx: vecIdx, fieldIdx: 0 });
-            if (!ctx.fast) fctx.body.push({ op: "f64.convert_i32_s" });
-          },
-          () => {
-            // [externref] → __extern_length (genuine host receiver / real JS array)
-            const lengthFuncIdx = ensureLateImport(ctx, "__extern_length", [{ kind: "externref" }], [{ kind: "f64" }]);
-            flushLateImportShifts(ctx, fctx);
-            if (lengthFuncIdx !== undefined) {
-              fctx.body.push({ op: "call", funcIdx: lengthFuncIdx });
-              if (ctx.fast) fctx.body.push({ op: "i32.trunc_sat_f64_s" });
-            } else {
-              fctx.body.push({ op: "drop" });
-              fctx.body.push({ op: ctx.fast ? "i32.const" : "f64.const", value: 0 });
-            }
-          },
-        );
-        return lenType;
-      }
-    }
-    // Shape-inferred array-like: obj.length → struct.get vec field 0
-    if (ts.isIdentifier(expr.expression)) {
-      const shapeInfo = ctx.shapeMap.get(expr.expression.text);
-      if (shapeInfo) {
-        compileExpression(ctx, fctx, expr.expression);
-        fctx.body.push({ op: "struct.get", typeIdx: shapeInfo.vecTypeIdx, fieldIdx: 0 });
-        if (!ctx.fast) fctx.body.push({ op: "f64.convert_i32_s" });
-        return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-      }
-    }
-    // Check the actual local type (may differ from TS type, e.g. arguments vec struct)
-    if (ts.isIdentifier(expr.expression)) {
-      const localIdx = fctx.localMap.get(expr.expression.text);
-      if (localIdx !== undefined) {
-        const localType =
-          localIdx < fctx.params.length
-            ? fctx.params[localIdx]!.type
-            : fctx.locals[localIdx - fctx.params.length]?.type;
-        // Vec struct ref local (e.g. `arguments` object) — struct.get field 0 (length)
-        // Note: for externref locals (e.g. `obj: any` in filter callbacks), we fall through
-        // to the generic externref path below (line ~1731) which uses multi-struct dispatch
-        // (ref.test → ref.cast → struct.get) to read the WasmGC struct field directly.
-        // Calling __extern_length on an externref-wrapped WasmGC struct returns 0 because
-        // obj.length is undefined on opaque externref objects in V8.
-        if ((localType?.kind === "ref" || localType?.kind === "ref_null") && localType.typeIdx !== undefined) {
-          const vecTypeIdx = (localType as { typeIdx: number }).typeIdx;
-          const typeDef = ctx.mod.types[vecTypeIdx];
-          // Plain vec ({length, data}) OR a `$__ta_view` ({length, buf, byteOffset},
-          // #3054 B1) — both keep the ELEMENT count at field 0.
-          if (
-            typeDef?.kind === "struct" &&
-            typeDef.fields[0]?.name === "length" &&
-            (typeDef.fields[1]?.name === "data" || isTaViewTypeIdx(ctx, vecTypeIdx))
-          ) {
-            if (isTaViewTypeIdx(ctx, vecTypeIdx)) {
-              // (#3054 C) A `$__ta_view` over a resizable buffer is auto-length —
-              // derive the CURRENT element count (field0 == -1 sentinel → live
-              // buf.length/elemSize) so `a.length` reflects a `rab.resize()`. A
-              // fixed view reads field0 directly (byte-identical to pre-C).
-              pushTaViewEffectiveLen(ctx, fctx, localIdx, vecTypeIdx);
-            } else {
-              fctx.body.push({ op: "local.get", index: localIdx });
-              fctx.body.push({ op: "struct.get", typeIdx: vecTypeIdx, fieldIdx: 0 });
-            }
-            if (!ctx.fast) fctx.body.push({ op: "f64.convert_i32_s" });
-            return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-          }
-        }
-      }
-    }
-    const objWasmType = resolveWasmType(ctx, objType);
-    if (objWasmType.kind === "ref" || objWasmType.kind === "ref_null") {
-      const vecTypeIdx = (objWasmType as { typeIdx: number }).typeIdx;
-      const typeDef = ctx.mod.types[vecTypeIdx];
-      if (typeDef?.kind === "struct" && typeDef.fields[1]?.name === "data") {
-        const exprResult = compileExpression(ctx, fctx, expr.expression);
-        // If the compiled expression returned externref (e.g. `x as any[]`), the TS type
-        // annotation doesn't guarantee the runtime struct type. Use multi-struct dispatch:
-        // try ref.test for the expected vec type first (struct.get field 0 gives the length),
-        // falling back to __extern_length for genuine host objects (real JS arrays).
-        // This avoids: (1) unguarded struct.get on externref (Wasm validation error), and
-        // (2) __extern_length returning 0 for WasmGC structs (obj.length is undefined on
-        // externref-wrapped WasmGC objects in V8).
-        if (exprResult?.kind === "externref") {
-          const extTmpIdx = allocLocal(fctx, `__len_ext_${fctx.locals.length}`, { kind: "externref" });
-          fctx.body.push({ op: "local.set", index: extTmpIdx });
-          const anyTmpIdx = allocLocal(fctx, `__len_any_${fctx.locals.length}`, { kind: "anyref" });
-          fctx.body.push({ op: "local.get", index: extTmpIdx });
-          fctx.body.push({ op: "any.convert_extern" });
-          fctx.body.push({ op: "local.set", index: anyTmpIdx });
-          const lenType = ctx.fast ? { kind: "i32" as const } : { kind: "f64" as const };
-          const lenTmp2 = allocLocal(fctx, `__len_val_${fctx.locals.length}`, lenType);
-          const lengthFuncIdx = ensureLateImport(ctx, "__extern_length", [{ kind: "externref" }], [{ kind: "f64" }]);
-          flushLateImportShifts(ctx, fctx);
-          const fallbackInstrs2: Instr[] =
-            lengthFuncIdx !== undefined
-              ? [
-                  { op: "local.get", index: extTmpIdx },
-                  { op: "call", funcIdx: lengthFuncIdx },
-                  ...(ctx.fast ? ([{ op: "i32.trunc_sat_f64_s" }] satisfies Instr[]) : []),
-                  { op: "local.set", index: lenTmp2 },
-                ]
-              : [
-                  { op: ctx.fast ? "i32.const" : "f64.const", value: 0 },
-                  { op: "local.set", index: lenTmp2 },
-                ];
-          fctx.body.push({ op: "local.get", index: anyTmpIdx });
-          fctx.body.push({ op: "ref.test", typeIdx: vecTypeIdx });
-          fctx.body.push({
-            op: "if",
-            blockType: { kind: "empty" },
-            then: [
-              { op: "local.get", index: anyTmpIdx },
-              { op: "ref.cast", typeIdx: vecTypeIdx },
-              { op: "struct.get", typeIdx: vecTypeIdx, fieldIdx: 0 },
-              ...(ctx.fast ? [] : ([{ op: "f64.convert_i32_s" }] satisfies Instr[])),
-              { op: "local.set", index: lenTmp2 },
-            ],
-            else: fallbackInstrs2,
-          });
-          fctx.body.push({ op: "local.get", index: lenTmp2 });
-          return lenType;
-        }
-        // Guard: the TS type might not match the runtime struct type.
-        // If the compiled expression returned a different ref type, use ref.test
-        // to verify before struct.get, falling back to 0.
-        if (
-          exprResult &&
-          (exprResult.kind === "ref" || exprResult.kind === "ref_null") &&
-          (exprResult as any).typeIdx !== vecTypeIdx
-        ) {
-          const lenTmp = allocLocal(fctx, `__len_tmp_${fctx.locals.length}`, { kind: "anyref" });
-          fctx.body.push({ op: "local.set", index: lenTmp });
-          fctx.body.push({ op: "local.get", index: lenTmp });
-          fctx.body.push({ op: "ref.test", typeIdx: vecTypeIdx });
-          const lenResult = ctx.fast ? { kind: "i32" as const } : { kind: "f64" as const };
-          fctx.body.push({
-            op: "if",
-            blockType: { kind: "val" as const, type: lenResult },
-            then: [
-              { op: "local.get", index: lenTmp },
-              { op: "ref.cast", typeIdx: vecTypeIdx },
-              { op: "struct.get", typeIdx: vecTypeIdx, fieldIdx: 0 },
-              ...(ctx.fast ? [] : ([{ op: "f64.convert_i32_s" }] satisfies Instr[])),
-            ],
-            else: [{ op: ctx.fast ? "i32.const" : "f64.const", value: 0 }],
-          });
-          return lenResult;
-        }
-        fctx.body.push({ op: "struct.get", typeIdx: vecTypeIdx, fieldIdx: 0 }); // get length from vec
-        if (!ctx.fast) fctx.body.push({ op: "f64.convert_i32_s" });
-        return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-      }
-    }
-    // Fallback: compile the expression and check the actual wasm return type
-    // This handles cases like strings.raw.length where TS doesn't know the type
-    {
-      // #1919 — transactional try-lower: keep the compiled receiver + struct.get
-      // when it lowers to a length-prefixed vec; otherwise roll back the body AND
-      // any locals / late imports / errors the compile leaked.
-      const snap = snapshotSpeculative(ctx, fctx);
-      const exprType = compileExpression(ctx, fctx, expr.expression);
-      if (
-        exprType &&
-        (exprType.kind === "ref" || exprType.kind === "ref_null") &&
-        (exprType as { typeIdx: number }).typeIdx !== undefined
-      ) {
-        const vecTypeIdx = (exprType as { typeIdx: number }).typeIdx;
-        const typeDef = ctx.mod.types[vecTypeIdx];
-        if (typeDef?.kind === "struct" && typeDef.fields[0]?.name === "length" && typeDef.fields[1]?.name === "data") {
-          fctx.body.push({ op: "struct.get", typeIdx: vecTypeIdx, fieldIdx: 0 });
-          if (!ctx.fast) fctx.body.push({ op: "f64.convert_i32_s" });
-          return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-        }
-      }
-      // (#2580 M2 slice 1) `.length` on a statically-`any`/`unknown` receiver in
-      // HOST mode, where the compiled receiver is NOT a length-bearing vec struct
-      // above. The origin path coerces the read to NUMERIC, turning a plain
-      // object's ABSENT `length` (spec `undefined`, §OrdinaryGet) into `0` / a
-      // bogus `typeof "boolean"` (`var obj={}; obj.length===undefined` → false —
-      // the #2580 headline bug). Route through the M2 tag/null-aware reader
-      // (`emitDynGet`), which returns a UNIFORM externref: a boxed number for a vec
-      // length / closure arity / null-undefined receiver (matching origin's prior
-      // numeric value — the #1894-eject Cluster A/B classes), JS `undefined` for a
-      // genuine non-null host object's absent property (the canary). The reader's
-      // receiver-kind dispatch (`__extern_is_undefined` → 0, `ref.test $vec` →
-      // field-0, `ref.test $closure` → 0, else `__extern_get`) is what a bare
-      // `__extern_get` could not do — the M1 over-broad arm's failure. Gated
-      // strictly on a static `any`/`unknown` receiver, host mode; the typed
-      // `.length` hot-path is byte-identical (handled + returned above).
-      // (#2580 M2 s1) DECLINE inside an async function/generator body. The
-      // async state machine (#1042 CPS lowering) can leave a destructuring rest /
-      // setter-captured local in a state where a speculative `compileExpression`
-      // recompile resolves a STALE value (the #2602-class desync; surfaces for the
-      // for-await array-rest `.length` reads incl. the setter-property variant).
-      // Origin reads those correctly via its own non-speculative path, so DECLINE
-      // here → fall through to origin (all 8 for-await rest `.length` tests stay
-      // green). The #2580 canary + Cluster A reads are NOT inside async functions,
-      // so they still take the reader. Walk to the nearest function-like ancestor
-      // and check the `async` modifier.
-      let inAsyncFn = false;
-      for (let p: ts.Node | undefined = expr.parent; p; p = p.parent) {
-        if (
-          ts.isFunctionDeclaration(p) ||
-          ts.isFunctionExpression(p) ||
-          ts.isArrowFunction(p) ||
-          ts.isMethodDeclaration(p)
-        ) {
-          inAsyncFn = ts.getModifiers(p)?.some((m) => m.kind === ts.SyntaxKind.AsyncKeyword) ?? false;
-          break;
-        }
-      }
-      if (
-        !ctx.standalone &&
-        !inAsyncFn &&
-        (objType.flags & (ts.TypeFlags.Any | ts.TypeFlags.Unknown)) !== 0 &&
-        exprType
-      ) {
-        if (exprType.kind !== "externref") {
-          coerceType(ctx, fctx, exprType, { kind: "externref" });
-        }
-        if (emitDynGet(ctx, fctx, "length")) {
-          return { kind: "externref" };
-        }
-        // emitDynGet bailed (no runtime) — roll back; the legacy paths recompile.
-        rollbackSpeculative(ctx, fctx, snap);
-      } else {
-        // (#1919) Undo the compiled expression if it didn't match — transactional
-        // rollback (body + locals + late imports + errors), not a bare truncate.
-        rollbackSpeculative(ctx, fctx, snap);
-      }
-    }
-    // #1472 Phase B Blocker B Slice 2 — standalone `.length` on an `any`/unknown
-    // receiver. None of the vec fast-paths matched, so the receiver is an opaque
-    // externref at runtime (e.g. the $ObjVec result of `Object.keys(o)` stored
-    // in an `any`). In standalone, `__extern_length` is the native $ObjVec
-    // reader (Blocker B Slice 1), so routing here keeps `.length` host-free and
-    // correct instead of falling through to `__extern_get("length")` (which the
-    // native `__extern_get` would mis-handle by casting "length" → key lookup,
-    // yielding 0). JS-host mode is unchanged (this gate is standalone-only; the
-    // host path's generic `__extern_get("length")` already works there).
-    if (ctx.standalone) {
-      const isAnyOrUnknown = (objType.flags & (ts.TypeFlags.Any | ts.TypeFlags.Unknown)) !== 0;
-      if (isAnyOrUnknown) {
-        const exprResult = compileExpression(ctx, fctx, expr.expression);
-        if (exprResult) {
-          if (exprResult.kind !== "externref") {
-            coerceType(ctx, fctx, exprResult, { kind: "externref" });
-          }
-          const lenFn = ensureLateImport(ctx, "__extern_length", [{ kind: "externref" }], [{ kind: "f64" }]);
-          flushLateImportShifts(ctx, fctx);
-          // (#2576, extends #2187) The `any`/unknown value may actually be a
-          // native `$AnyString` (object string-value read, generator yield read,
-          // catch binding, indexed element read). `__extern_length` reads
-          // $ObjVec/array length and returns 0 for a bare string, so guard with
-          // `ref.test $AnyString` first: a string hit reads `$AnyString.len`
-          // (field 0); the miss falls to the existing `__extern_length`
-          // array/$ObjVec reader. Always computes an i32 length, converted to f64
-          // once below in !fast mode. sd-3's `receiverIsNativeStringValType` arm
-          // (above) already handles the bare-identifier-with-string-ref-local
-          // case; this covers the opaque-externref cases it cannot see statically.
-          if (ctx.nativeStrings && ctx.anyStrTypeIdx >= 0) {
-            emitGuardedNativeStringLength(ctx, fctx, (recvExternLocal) =>
-              lenFn !== undefined
-                ? [
-                    { op: "local.get", index: recvExternLocal },
-                    { op: "call", funcIdx: lenFn },
-                    { op: "i32.trunc_sat_f64_s" },
-                  ]
-                : [{ op: "i32.const", value: 0 }],
-            );
-            if (!ctx.fast) fctx.body.push({ op: "f64.convert_i32_s" });
-            return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-          }
-          if (lenFn !== undefined) {
-            fctx.body.push({ op: "call", funcIdx: lenFn });
-            if (ctx.fast) fctx.body.push({ op: "i32.trunc_sat_f64_s" });
-            return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-          }
-          // No helper available — drop and yield 0.
-          fctx.body.push({ op: "drop" });
-          fctx.body.push({ op: ctx.fast ? "i32.const" : "f64.const", value: 0 });
-          return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-        }
-      }
-    }
-  }
-
-  // Handle .raw on tagged template strings arrays (template vec struct)
-  // The strings parameter is typed as a base vec, but at runtime it's a
-  // template vec (subtype with an extra raw field). We ref.cast to the
-  // template vec type and then struct.get field 2.
-  if (propName === "raw" && ctx.templateVecTypeIdx >= 0) {
-    const templateVecTypeIdx = ctx.templateVecTypeIdx;
-    // Check if the object is a vec-like type (base vec or template vec)
-    let isVecLike = false;
-    if (ts.isIdentifier(expr.expression)) {
-      const localIdx = fctx.localMap.get(expr.expression.text);
-      if (localIdx !== undefined) {
-        const localType =
-          localIdx < fctx.params.length
-            ? fctx.params[localIdx]!.type
-            : fctx.locals[localIdx - fctx.params.length]?.type;
-        if ((localType?.kind === "ref" || localType?.kind === "ref_null") && localType.typeIdx !== undefined) {
-          const typeIdx = (localType as { typeIdx: number }).typeIdx;
-          const typeDef = ctx.mod.types[typeIdx];
-          if (
-            typeDef?.kind === "struct" &&
-            typeDef.fields[0]?.name === "length" &&
-            typeDef.fields[1]?.name === "data"
-          ) {
-            isVecLike = true;
-          }
-        }
-      }
-    }
-    if (!isVecLike) {
-      const objWasmType = resolveWasmType(ctx, objType);
-      if (objWasmType.kind === "ref" || objWasmType.kind === "ref_null") {
-        const typeIdx = (objWasmType as { typeIdx: number }).typeIdx;
-        const typeDef = ctx.mod.types[typeIdx];
-        if (typeDef?.kind === "struct" && typeDef.fields[0]?.name === "length" && typeDef.fields[1]?.name === "data") {
-          isVecLike = true;
-        }
-      }
-    }
-    if (isVecLike) {
-      // Compile the object expression, cast to template vec, and get raw field
-      // Guard with ref.test to avoid illegal cast trap if the runtime type
-      // is a base vec (not a template vec with the extra raw field).
-      compileExpression(ctx, fctx, expr.expression);
-      const baseVecTypeIdx = getOrRegisterVecType(ctx, "externref", { kind: "externref" });
-      const rawTmp = allocLocal(fctx, `__raw_tmp_${fctx.locals.length}`, { kind: "ref_null", typeIdx: baseVecTypeIdx });
-      const rawObj = allocLocal(fctx, `__raw_obj_${fctx.locals.length}`, { kind: "anyref" });
-      fctx.body.push({ op: "local.set", index: rawObj });
-      fctx.body.push({ op: "local.get", index: rawObj });
-      fctx.body.push({ op: "ref.test", typeIdx: templateVecTypeIdx });
-      fctx.body.push({
-        op: "if",
-        blockType: { kind: "empty" },
-        then: [
-          { op: "local.get", index: rawObj },
-          { op: "ref.cast", typeIdx: templateVecTypeIdx },
-          { op: "struct.get", typeIdx: templateVecTypeIdx, fieldIdx: 2 },
-          { op: "local.set", index: rawTmp },
-        ],
-        else: [
-          // Not a template vec — return null (no raw field available)
-          { op: "ref.null", typeIdx: baseVecTypeIdx },
-          { op: "local.set", index: rawTmp },
-        ],
-      });
-      fctx.body.push({ op: "local.get", index: rawTmp });
-      return { kind: "ref_null", typeIdx: baseVecTypeIdx };
-    }
-  }
-
-  // Handle Math constants
-  if (ts.isIdentifier(expr.expression) && expr.expression.text === "Math") {
-    const mathConstants: Record<string, number> = {
-      PI: Math.PI,
-      E: Math.E,
-      LN2: Math.LN2,
-      LN10: Math.LN10,
-      SQRT2: Math.SQRT2,
-      SQRT1_2: Math.SQRT1_2,
-      LOG2E: Math.LOG2E,
-      LOG10E: Math.LOG10E,
-    };
-    if (propName in mathConstants) {
-      fctx.body.push({ op: "f64.const", value: mathConstants[propName]! });
-      return { kind: "f64" };
-    }
-  }
-
-  // Handle Number constants
-  if (ts.isIdentifier(expr.expression) && expr.expression.text === "Number") {
-    const numberConstants: Record<string, number> = {
-      EPSILON: Number.EPSILON,
-      MAX_SAFE_INTEGER: Number.MAX_SAFE_INTEGER,
-      MIN_SAFE_INTEGER: Number.MIN_SAFE_INTEGER,
-      MAX_VALUE: Number.MAX_VALUE,
-      MIN_VALUE: Number.MIN_VALUE,
-      POSITIVE_INFINITY: Infinity,
-      NEGATIVE_INFINITY: -Infinity,
-      NaN: NaN,
-    };
-    if (propName in numberConstants) {
-      fctx.body.push({ op: "f64.const", value: numberConstants[propName]! });
-      return { kind: "f64" };
-    }
-  }
-
-  // (#2595) `<TypedArrayName>.BYTES_PER_ELEMENT` — static element byte width
-  // (§23.2.6.x). Statically known per constructor name, so emit it as a
-  // constant. Standalone otherwise reaches `reportUnsupportedStandaloneBuiltinValueRead`
-  // (the generic builtin-static-value-read refusal); host mode reads the same
-  // constant via the host import, so folding it here is observationally
-  // identical and works in both modes. Skip when the name is shadowed by a local.
-  if (
-    propName === "BYTES_PER_ELEMENT" &&
-    ts.isIdentifier(expr.expression) &&
-    expr.expression.text in TYPED_ARRAY_BYTES_PER_ELEMENT
-  ) {
-    const builtinName = expr.expression.text;
-    const isShadowed = fctx.localMap.has(builtinName) || (fctx.boxedCaptures?.has(builtinName) ?? false);
-    if (!isShadowed) {
-      const bytes = TYPED_ARRAY_BYTES_PER_ELEMENT[builtinName]!;
-      fctx.body.push({ op: ctx.fast ? "i32.const" : "f64.const", value: bytes });
-      return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-    }
-  }
-
-  // (#2861) `<Ctor>.length` (declared arity) / `<Ctor>.name` (ctor name string)
-  // — a built-in constructor's own function properties. Statically known per ctor
-  // name, so emit as a constant. Standalone otherwise reaches
-  // `reportUnsupportedStandaloneBuiltinValueRead` (the generic builtin-static-value
-  // -read refusal); host mode reads the same value via `__get_builtin` and returns
-  // BEFORE this point, so folding here is observationally identical and never
-  // fires in host mode for a ctor. Namespaces (Math/JSON/Reflect/Atomics) are not
-  // in BUILTIN_CTOR_ARITY (their `.length`/`.name` are undefined), so they keep
-  // refusing. Skip when the name is shadowed by a local.
-  if (
-    (propName === "length" || propName === "name") &&
-    ts.isIdentifier(expr.expression) &&
-    expr.expression.text in BUILTIN_CTOR_ARITY
-  ) {
-    const builtinName = expr.expression.text;
-    const isShadowed = fctx.localMap.has(builtinName) || (fctx.boxedCaptures?.has(builtinName) ?? false);
-    if (!isShadowed) {
-      if (propName === "length") {
-        const arity = BUILTIN_CTOR_ARITY[builtinName]!;
-        fctx.body.push({ op: ctx.fast ? "i32.const" : "f64.const", value: arity });
-        return ctx.fast ? { kind: "i32" } : { kind: "f64" };
-      }
-      // `<Ctor>.name` === "<Ctor>" for every standard builtin constructor.
-      addStringConstantGlobal(ctx, builtinName);
-      fctx.body.push(...stringConstantExternrefInstrs(ctx, builtinName));
-      return { kind: "externref" };
-    }
-  }
-
-  // Handle Symbol.iterator, Symbol.hasInstance, etc. → constant i32
-  if (ts.isIdentifier(expr.expression) && expr.expression.text === "Symbol") {
-    const symId = getWellKnownSymbolId(propName);
-    if (symId !== undefined) {
-      fctx.body.push({ op: "i32.const", value: symId });
-      return { kind: "i32" };
-    }
-  }
-
-  // (#1467) `sym.description` — Symbol.prototype.description accessor.
-  // When the LHS is a Symbol primitive (or Symbol-wrapper object), read the
-  // host's Symbol.prototype.description accessor via `__symbol_description`.
-  // This handles three test262 buckets:
-  //   • Symbol('x').description === 'x'
-  //   • Symbol().description === undefined
-  //   • Symbol.prototype.description.call(wrapperObj) → unwraps the wrapper
-  // Generic __extern_get works for plain JS hosts but bypasses the spec
-  // accessor (which V8 implements specially), so we route directly.
-  if (propName === "description" && (objType.flags & ts.TypeFlags.ESSymbolLike) !== 0) {
-    // (#2163) No-JS-host mode: the symbol is a bare i32 id and there is no host
-    // accessor — read the description from the native id→string side table
-    // (populated by `compileSymbolCall`). A null slot / out-of-range id reads as
-    // `undefined`, matching `Symbol().description === undefined`.
-    if (noJsHost(ctx)) {
-      const recvType = compileExpression(ctx, fctx, expr.expression, { kind: "i32" });
-      if (recvType && recvType.kind !== "i32") {
-        coerceType(ctx, fctx, recvType, { kind: "i32" });
-      }
-      emitSymbolDescLoad(ctx, fctx);
-      // Result is `ref_null $AnyString` — a native string (or null⇒undefined).
-      return { kind: "ref_null", typeIdx: ctx.anyStrTypeIdx };
-    }
-    const symDescIdx = ensureLateImport(ctx, "__symbol_description", [{ kind: "externref" }], [{ kind: "externref" }]);
-    if (symDescIdx !== undefined) {
-      const recvType = compileExpression(ctx, fctx, expr.expression, { kind: "externref" });
-      if (recvType && recvType.kind !== "externref") {
-        coerceType(ctx, fctx, recvType, { kind: "externref" });
-      }
-      flushLateImportShifts(ctx, fctx);
-      fctx.body.push({ op: "call", funcIdx: symDescIdx });
-      return { kind: "externref" };
-    }
-  }
-
-  // #1910 R4 — String-wrapper `.length` in standalone. `new String("ab")` builds
-  // a `$Object` wrapper carrying its [[StringData]] native string in the reserved
-  // FLAG_INTERNAL slot (#1910 S2). `.length` is a String-exotic own property whose
-  // value is the underlying string's length (§22.1.4.1). Recover the slot string
-  // via `__to_primitive(recv, "string")` (reads the slot first, §7.1.1.1), then
-  // read `$AnyString.len` (field 0). Standalone only — host mode keeps the wrapper
-  // host-object machinery and its own `.length` reader.
-  if (ctx.standalone && isStringWrapperType(objType) && propName === "length" && ctx.anyStrTypeIdx >= 0) {
-    ensureObjectRuntime(ctx);
-    const toPrimIdx = ctx.funcMap.get("__to_primitive");
-    if (toPrimIdx !== undefined) {
-      compileExpression(ctx, fctx, expr.expression, { kind: "externref" });
-      addStringConstantGlobal(ctx, "string");
-      fctx.body.push(...stringConstantExternrefInstrs(ctx, "string"));
-      fctx.body.push({ op: "call", funcIdx: toPrimIdx });
-      // __to_primitive returns the [[StringData]] string as externref; coerce to
-      // $AnyString and read its `len` field.
-      coerceType(ctx, fctx, { kind: "externref" }, { kind: "ref_null", typeIdx: ctx.anyStrTypeIdx });
-      fctx.body.push({ op: "struct.get", typeIdx: ctx.anyStrTypeIdx, fieldIdx: 0 });
-      return { kind: "i32" };
-    }
-  }
-
-  // Handle string.length
-  // (#2187) Also fire for an `any`-typed identifier whose compiled local ValType
-  // is a native-string ref (e.g. a for-of var from a string-yielding generator):
-  // the static type lost the string info, but at the VALUE level it IS a string.
-  if ((isStringType(objType) || receiverIsNativeStringValType(ctx, fctx, expr.expression)) && propName === "length") {
-    const recvType = compileExpression(ctx, fctx, expr.expression);
-    if (ctx.nativeStrings && ctx.anyStrTypeIdx >= 0) {
-      // The receiver must be a `$AnyString` ref before reading its `len`
-      // field. Some string producers (e.g. the native Error `.name`/`.message`
-      // reader, #1104/#1797) hand back an `externref`; coerce it to the GC
-      // string ref first, otherwise `struct.get $AnyString` validates against
-      // an externref operand → invalid Wasm (#1797).
-      if (recvType && recvType.kind === "externref") {
-        coerceType(ctx, fctx, recvType, { kind: "ref_null", typeIdx: ctx.anyStrTypeIdx });
-      }
-      // len is field 0 of $AnyString — works for both FlatString and ConsString
-      fctx.body.push({ op: "struct.get", typeIdx: ctx.anyStrTypeIdx, fieldIdx: 0 });
-      return { kind: "i32" };
-    }
-    const funcIdx = ctx.jsStringImports.get("length");
-    if (funcIdx !== undefined) {
-      fctx.body.push({ op: "call", funcIdx });
-      return { kind: "i32" };
-    }
-  }
-
-  // Handle IteratorResult property access: .value and .done
-  if (isIteratorResultType(objType) || isGeneratorIteratorResultLike(ctx, objType, propName)) {
-    const nativeResult = tryCompileNativeGeneratorResultProperty(ctx, fctx, expr.expression, propName);
-    if (nativeResult !== undefined) return nativeResult;
-    if (propName === "value") {
-      compileExpression(ctx, fctx, expr.expression);
-      // Check the expected value type from the IteratorResult<T>. NOTE (#2030):
-      // an exhausted result's `.value` is `undefined`; the f64 fast path below
-      // runs `Number(undefined)` → NaN, so a string context of the
-      // value-after-done prints "NaN". Making that survive as "undefined"
-      // requires the value-buffer representation work tracked by #2035 and is
-      // intentionally NOT changed here — routing `.value` through externref
-      // breaks numeric consumers (illegal cast on the raw-f64 iteration path).
-      const valueType = getIteratorResultValueType(ctx, objType);
-      if (valueType && valueType.kind === "f64") {
-        const funcIdx = ctx.funcMap.get("__gen_result_value_f64");
-        if (funcIdx !== undefined) {
-          fctx.body.push({ op: "call", funcIdx });
-          return { kind: "f64" };
-        }
-      }
-      const funcIdx = ctx.funcMap.get("__gen_result_value");
-      if (funcIdx !== undefined) {
-        fctx.body.push({ op: "call", funcIdx });
-        return { kind: "externref" };
-      }
-    }
-    if (propName === "done") {
-      compileExpression(ctx, fctx, expr.expression);
-      const funcIdx = ctx.funcMap.get("__gen_result_done");
-      if (funcIdx !== undefined) {
-        fctx.body.push({ op: "call", funcIdx });
-        // #2030: `.done` is a boolean — brand it so string contexts render
-        // "true"/"false" rather than the raw i32 "1"/"0".
-        return { kind: "i32", boolean: true };
-      }
-    }
-  }
-
-  // Handle externref property access
-  if (isExternalDeclaredClass(objType, ctx.checker)) {
-    const externResult = compileExternPropertyGet(ctx, fctx, expr, objType, propName);
-    if (externResult !== null) return externResult;
-    // Fall through to dynamic fallback if import is missing
+  {
+    const __r = tryStringLengthIteratorAndExternClassReads(ctx, fctx, expr, propName, objType);
+    if (__r !== PA_FALLTHROUGH) return __r;
   }
 
   // Handle getter accessor on user-defined classes
@@ -4387,7 +3487,7 @@ export function compilePropertyAccess(
   return { kind: "externref" };
 }
 
-function compileExternPropertyGet(
+export function compileExternPropertyGet(
   ctx: CodegenContext,
   fctx: FunctionContext,
   expr: ts.PropertyAccessExpression,
@@ -4870,7 +3970,7 @@ export function emitThisReceiverGuardConvert(
  *   - for a `.member` access, the registered vec types are NOT added (a bare
  *     `this.member` on an untyped receiver stays on the host path).
  */
-function thisReceiverGuardTargets(
+export function thisReceiverGuardTargets(
   ctx: CodegenContext,
   fctx: FunctionContext,
   objExpr: ts.Expression,
